requirements.txt
google-generativeai
pandas
python-dotenv
nltk
scikit-learn
seaborn
matplotlib
joblib
streamlit
plotly